Initial support for trunkdev
The 'trunkdev' is a virtual DAHDI device driver that exposes the
multiplexed E1 frames on a /dev/dahdi/trunkdev device.  This is useful
for all kinds of simulated setup, as well as in  the context of
using osmo-e1d for OCTOI (Osmocom TDMoIP).

In order for the dahdi_trunkdev utility to be built, you will need
to have installed dahdi header files of dahdi-linux with trunkdev
support, currently located at
https://gitea.osmocom.org/retronetworking/dahdi-linux/src/branch/laforge/trunkdev

Change-Id: Ie05100c80fd978e7ac6f5a028e0ef6c828e0bcc3